test/python/requirements.txt
duckdb
pytest
spherely
geoarrow-pyarrow
